requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==1.4.49
psycopg2-binary==2.9.9
azure-storage-blob==12.19.0
azure-core==1.29.5
aiohttp==3.9.1
aiofiles==23.2.1
python-dotenv==1.0.0
typing-extensions>=4.8.0
multidict>=4.5.0
yarl>=1.6.0
frozenlist>=1.1.1
attrs>=17.3.0 
PyPDF2>=3.0.0
Pillow>=9.0.0
python-docx>=0.8.11
pytesseract>=0.3.0
pymupdf>=1.26.0
fitz==0.0.1.dev2
